MP-166 :truck: Rename: commons 패키지의 이름을 common으로 변경
* commons와 common 모두 좋기는 하나 패키지 뒤에 s를 붙이는 경우는 반드시 그래야만 할 때(enums, supers 등)로 제한하여 패키지 네이밍의 통일성 확보
