+++ src/main/java/kr/modusplant/domains/common/persistence/repository/supers/CreatedAtAndLastModifiedAtRepository.java
@@ -1,4 +1,4 @@
-package kr.modusplant.domains.commons.persistence.repository.supers;
+package kr.modusplant.domains.common.persistence.repository.supers;
 
 public interface CreatedAtAndLastModifiedAtRepository<T> extends CreatedAtRepository<T>, LastModifiedAtRepository<T> {
 }
