+++ src/main/java/kr/modusplant/domains/common/persistence/repository/supers/LastModifiedAtRepository.java
@@ -1,4 +1,4 @@
-package kr.modusplant.domains.commons.persistence.repository.supers;
+package kr.modusplant.domains.common.persistence.repository.supers;
 
 import java.time.LocalDateTime;
 import java.util.List;
