+++ src/main/java/kr/modusplant/domains/common/vo/Reference.java
@@ -1,4 +1,4 @@
-package kr.modusplant.domains.commons.vo;
+package kr.modusplant.domains.common.vo;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
