+++ src/main/java/kr/modusplant/domains/member/persistence/repository/SiteMemberTermRepository.java
@@ -2,7 +2,7 @@ package kr.modusplant.domains.member.persistence.repository;
 
 import kr.modusplant.domains.member.persistence.entity.SiteMemberTermEntity;
 import kr.modusplant.domains.member.persistence.repository.supers.SiteMemberUuidPrimaryKeyRepository;
-import kr.modusplant.domains.commons.persistence.repository.supers.LastModifiedAtRepository;
+import kr.modusplant.domains.common.persistence.repository.supers.LastModifiedAtRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
