+++ src/main/java/kr/modusplant/modules/jwt/persistence/repository/RefreshTokenRepository.java
@@ -1,7 +1,7 @@
 package kr.modusplant.modules.jwt.persistence.repository;
 
 import kr.modusplant.domains.member.persistence.entity.SiteMemberEntity;
-import kr.modusplant.domains.commons.persistence.repository.supers.UuidPrimaryKeyRepository;
+import kr.modusplant.domains.common.persistence.repository.supers.UuidPrimaryKeyRepository;
 import kr.modusplant.modules.jwt.persistence.entity.RefreshTokenEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
